requirements.txt
bcrypt==4.0.1
black==23.1.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
markdown-it-py==2.1.0
mdurl==0.1.2
mypy==1.0.0
mypy-extensions==1.0.0
packaging==23.0
paramiko==3.0.0
pathspec==0.11.0
platformdirs==3.0.0
pycparser==2.21
Pygments==2.14.0
PyNaCl==1.5.0
python-dotenv==0.21.1
rich==13.3.1
tomli==2.0.1
types-paramiko==3.0.0.2
typing_extensions==4.4.0
